yardlistr: Generate Summary Plots from eBird Data

The yardlistr package loads your personal eBird data from a csv file, processes checklists from a user-specified list of locations and generates a variety of plots for each location.

See also

Author

Maintainer: Gleb G. Ebert dev@gl-eb.me (ORCID) [copyright holder]